use Select instead of Match; moved tests into existing test project
