+++ Functional.Tests/Utilities/Assertions/OptionTypeAssertions.cs
@@ -2,9 +2,8 @@
 using FluentAssertions;
 using FluentAssertions.Equivalency;
 using FluentAssertions.Execution;
-using Functional;
 
-namespace FluentAssertions
+namespace Functional.Tests.Utilities.Assertions
 {
 	/// <summary>
 	/// Defines assertions for <see cref="Option{TValue}"/> type.
